+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## 4.0.2
+
+This is just a maintenance release with a bunch of bug fixes.
+
+### Added
+
+- VLM: use airfoil polars with forced transition to get better alpha0 of airfoil
+
+### Fixed
+
+- VLM: sometimes airfoils were not straked
+
+
 ## 4.0.1
 
 This is just a maintenance release with a bunch of bug fixes.